+++ src/Domains/Identity/Taskify.Identity.UseCases/UserContextRoles/Delete/DeleteUserContextRoleHandler.cs
@@ -1,6 +1,5 @@
 ﻿namespace Taskify.Identity.UseCases.UserContextRoles.Delete;
 
-using Ardalis.GuardClauses;
 using Ardalis.Result;
 
 using Taskify.Identity.Core.UserContextRoleAggregate;
@@ -22,7 +21,8 @@ public class DeleteUserContextRoleHandler : ICommandHandler<DeleteUserContextRol
         CancellationToken cancellationToken)
     {
         var item = await _repository.GetByIdAsync(request.Id, cancellationToken);
-        Guard.Against.Null(item);
+        if (item == null)
+            return Result.NotFound("UserContextRole not found.");
         await _repository.DeleteAsync(item, cancellationToken);
         return Result.Success();
     }
